Added Crazy enemy and health bar to the UI
Moved all health related code for the player to PlayerData,

Health code that was in Actor was moved to Enemy.gd as it is no longer needed in the Player node.

Crazy enemy generates a random every .2 seconds, if the number returned is greater than 80 it will change direction.  If it has touched a wall it will wait .5 seconds before changing directions again.